Clean up interface for IDEs
